add proteins data
Added proteins data v1 from zenodo download, which is linked to the proteins package. Will need to update this to v1.1 when we have more information that we'd like to include, such as the 'description' field. Also added proteins to the test data generation. And added the URL to zenodo to the current_release.R, with other URLs.