improve retrying "path" step in generate_data.R
We have seen problems with enrichr failing after several
hours when running the "path" step. This results in no subset
files being created and requiring starting over. To better
handle this scenario we increase the number of subset files
so progress is saved and do not re-create subset files.
After the pathway files are created the subset files are now
deleted.