Update code to generate 20Q3 data
Added code to generate tables for P 53 Jeanne that will then allow a network graph to be made. Previous problem was needing to generate data before generating methods. A better solution is to generate a subset of the data, Render the methods, and then push the entire change to github. This code also changed how some of the master tables were made. I broke out two small functions, that are now called in the script. Part of the reason I did this was to be able to reuse some of the code in generating the test data set.